Handling of type A tags that don't answer to S(DESELECT)
In reSelect() procedure, after having send deact(sleep), the tag
is in idle mode, discovery is restarted and tag detected anew if still
present. Wait for this detection and do not exit with error.
When reSelect() procedure succeeds after restarting the discovery,
we might not be connected to the requested interface.
Check and update variables when exiting doConnect(),
update when exiting doReconnect().

Bug: 340223581
Test: manual
Change-Id: Icd238f9c9ac5d33e6bc1f3d608e903f23b5751e2