nfc: Fix a bug in RegisteredServicesCache
The bug is that the uid is not correctly set when creating a new DynamicSettings object from within the NFC stack (migration).
This can cause problems when the object is later used.

Also, removed uid checking on parsing which is redundant and will
help users upgrade to fix the issue.

Bug: 379683345
Change-Id: I72e977bb45ae7ed992e07d9592a1ce2a3e40fdfa
Test: Manual verification